location-source: Fix subclass resource leaks on start
The start_source() and stop_source() functions in GClueLocationSource
are chained up by all the start() and stop() method implementations in
GClueLocationSource subclasses to determine when to really start and
stop the source and allocate / free the needed resources.

Commit 0661107fe0 made start_source() return TRUE when the source was
already active, which caused the resources already reserved for the
physical source to be reserved again. The source class implementations
are typically designed to hold just one set of resources, so the
references to previously reserved resources are lost, that is they leak.
An example of this is the GSocketClient reference (priv->client) held by
GClueNMEASource.

Another visible (in the debug log) example of the extra resources are
callbacks connected to the "fix-*" signals of the physical sources (see
gclue_modem_gps_start() for example). With more than one client using
the same source, all clients receive the same location update N times,
where N is the number of clients using the source.

Make GClueLocationSource.start() and .stop() methods return enums
GClueLocationSourceStartResult and GClueLocationSourceStopResult which
gives subclasses more information on when to allocate resources on
starting and stopping.

The start_source() function now returns the value
GCLUE_LOCATION_SOURCE_START_RESULT_ALREADY_STARTED when the source is
already active. The subclass start() method implementations are changed
to return immediately when they get this value from the base class.

This effectively reverts the first diff in commit 0661107fe0
"location-source: Return source state from start_source and
stop_source".

The behaviour of stop() method subclass implementations remains the
same. Freeing of resources is always attempted, except when the value
GCLUE_LOCATION_SOURCE_STOP_RESULT_STILL_USED is returned from the base
class.