location: Fix gclue_location_set_heading_from_prev_location()
gclue_location_set_heading_from_prev_location() uses atan2 to convert
the movement vector coordinates to a bearing (positive degrees from true
north, clockwise) which is then used for the "heading" GClueLocation
property.

The implementation did not change the sign of the atan2 output and had a
wrong offset, resulting in random looking heading output.

Fix the implementation and improve the comments.