Define WiFi APs noise floor in a single place and fix its debug messages
WiFi APs noise floor was hardcoded in two different places in code, let's
add a common define for it for clarity.

While we are at it let's also fix the debug messages that are printed when
an AP is ignored due to its signal level being too low - the level was
being printed as an unsigned value so the usual (negative) signal levels
were printed as huge positive values.