service-client: Be strict with time and distance thresholds
The 'LocationUpdated' signal should be emitted when a new location
either has a timestamp that is newer by 'TimeThreshold' than the
previous signaled location, or if the distance of the new location is
larger than 'DistanceThreshold' from the previous signaled location.

The 'Location' property can be updated even if the location is below the
thresholds for signaling it, allowing clients to monitor location
updates which are below the signaling thresholds.

When a new location arrives, it is compared to the Location property to
check if time and distance thresholds are exceeded. However, as Location
can be updated without signaling, a client expecting LocationUpdated
signals every N seconds may never receive one, if updates are always
less than N seconds apart. The same can happen with the distance
threshold, if the location moves slowly enough.

Remember the signaled location and make time and distance comparisons
for new locations against it, so that clients get LocationUpdated
signals strictly within the requested time and distance thresholds.