service-manager: Replace agent wait timeout with a queue
When GeoClue is started with D-Bus activation, the agent is not started
yet and the starting client currently has to wait AGENT_WAIT_TIMEOUT
(=20 s) before being started.

Replace the agent wait timeout with a FIFO queue which is processed as
soon as the agent connects.

Handle the case where client vanishes before agent is connected.

Remove the now unused var priv->init_time.

Fixes #141.