Add static location source
There were many requests to provide a static location source for systems
which don't normally move but where the existing location sources provide
poor location estimate.
An example of such system would be a desktop PC without WiFi or 3G modem.

So let's add a source that reads location from "geolocation" file in
@sysconfdir@ (normally /etc) to cover this use case.
This file is constantly monitored for changes during Geoclue operation
and the reported static location is updated accordingly.

The geoclue(5) man page should be consulted for the format description of
this file.