Revert "modem-manager: Only disable the given capability in clear_caps()"
mm_modem_location_get_enabled() uses g_dbus_proxy_get_cached_property()
under the hood, so the returned value will not be correct if MM caps are
updated by more than one source close in time.

This can happen for example when a client using both MM GPS and 3G
sources quits. The GPS capability is cleared first, but the update is
not reflected in the cached 'enabled' property. The 3G cap is cleared
next, but it gets an incorrect value from mm_modem_location_get_enabled
and turns the GPS back on.

This reverts commit a9caa550df128bb04f28b3f3c94d909b0d140e15.