lib: simple: cancel on cancellable signal
When the cancellable is cancelled, the tasks sub-tasks
are cancelled, and chain up the cancellation error.

However, when the sub-tasks all finish successfully,
a location-updated signal handler is attached and the
init only finishes after that. If that is never triggered,
which is not guaranteed not to happen, cancelling the
init task has no effect and it is kept dormant.

For that, a signal handler is needed.

As the task now has multiple places that can return,
it also needs g_task_get_completed checks to avoid
multiple returns.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>