lib: simple: make priv->task a weak pointer
While through this pointer the one strong reference
to the task is controlled by unrefing, this pointer
itself gets no reference for itself. Thus, it limits the
ability to control the task without having access to the
GClue.Simple.

To make this relationship more clear, the g_clear_object
is made g_object_unref instead, and letting the weak
pointer handle setting the pointer to NULL.

The alternative would be to use a strong reference with
g_set_object. That, however, would introduce a reference
cycle between the task and the GClue.Simple.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>