lib: simple: use g_signal_connect_object for auto-disconnection
g_signal_connect_object creates a closure that is
destroyed when the object – in this case the GClue.Simple –
is destroyed. This ensures the signal handler is
automatically disconnected. That was manually done before
in _finalize for one of the two g_signal_connect users,
but not done for the other.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>