location: Ignore NMEA coordinates without degree digits
The coordinates for latitude and longitude in NMEA sentences are given
as strings in the format `ddmm.mm` where ds are the digits for degrees
and mm.mm is the float value for minutes. Degree values for latitudes
are between 1 to 2 digits (0 to 90) and between 1 to 3 digits (0 to 180)
for longitudes. Implementations seem to mostly output degrees
zero-padded with 2 digits for latitude and 3 digits with longitude.

We now reject coordinate strings without at least one character for
degree, so that e.g. "012.3456" is allowed, but "12.3456" is not.

Previously, coordinate strings without digit values were parsed, leading
to erroneous output.

Thanks to Benoit Maurin for reporting this.