location: Stricter NMEA coordinate string parsing
Add check for NULL coordinate and direction strings, remove duplicate
check for empty direction string.

Check coordinate string for chars which are not a digit or '.' and
reject such strings.

Check that the full minutes part of the string is consumed by
g_ascii_strtod(), reject the string if not. The degrees part is covered
by other checks on the coordinate string.