Fix AppArmor rules for DBus
Verified working on Ubuntu.

I've tried securing most of the calls, however because this service registers
objects and sends signals about those it must be in some way allowed to send
messages to anyone on the bus.
For this reason, you'll find that for non-sensitive object paths I have not
restricted the peer name.

LP: #2030951