wifi: Do not create a location without a BSS list
Create a new location only when a scan has finished and it contains a
non-empty set of BSS's:

Replace full source refresh with available accuracy level refresh when
there is no wifi interface on initialization, when the wifi source is
not set to active, and on wifi interface removal.

Don't refresh the source or available accuracy level on construction,
when there is no configured interface yet.

Return G_IO_ERROR_NOT_INITIALIZED on refresh_async() if the BSS list
from a scan is empty.